+++ librbr/include/core/rewards/saso_rewards_array.h
@@ -147,12 +147,12 @@ private:
 	/**
 	 * The minimum R-value.
 	 */
-	float Rmin;
+	double Rmin;
 
 	/**
 	 * The maximum R-value.
 	 */
-	float Rmax;
+	double Rmax;
 
 };
